+++ Xabbo.Core/Items/IFurni.cs
@@ -10,7 +10,7 @@ namespace Xabbo.Core
         /// <summary>
         /// Gets the owner of the furni's ID
         /// </summary>
-        int OwnerId { get; }
+        long OwnerId { get; }
         /// <summary>
         /// Gets the owner of the furni's name
         /// </summary>
